cloud_brain/requirements.txt

fastapi
uvicorn
supabase
python-dotenv
google-genai
twilio
python-multipart
